lxd provider: use right bind address for LXD
Prevents:

ERROR cannot find network interface "lxcbr0": route ip+net: no such network interface
ERROR invalid config: route ip+net: no such network interface

when the user doesn't have a lxcbr0 (i.e. doesn't have lxc1 installed).
