+++ container/lxd/lxd.go
@@ -7,7 +7,6 @@ package lxd
 
 import (
 	"fmt"
-	"os"
 
 	"github.com/juju/errors"
 	"github.com/juju/loggo"
@@ -267,19 +266,4 @@ func (manager *containerManager) createNetworkProfile(profile string, networkCon
 	}
 
 	return nil
-}
-
-// GetDefaultBridgeName returns the name of the default bridge for lxd.
-func GetDefaultBridgeName() (string, error) {
-	_, err := os.Lstat("/sys/class/net/lxdbr0/bridge")
-	if err == nil {
-		return "lxdbr0", nil
-	}
-
-	/* if it was some unknown error, return that */
-	if !os.IsNotExist(err) {
-		return "", err
-	}
-
-	return "lxcbr0", nil
 }
